[srm] CDPD-29124: MM2 should support global worker and connector configurations
MM2 only propagates the explicitly declared configurations and the replication flow specific configurations to the Connector configurations. This is important in case of a custom MetricsReporter or TopicFilter implementation. Similarly, only a handful of Connect worker configurations are allowed to be configured globally.
To support passing arbitrary configurations to the underlying Connectors, there should be a prefix "connectors." which gets propagated to the Connector configuration, with the prefix stripped. Similarly, to support any Connect worker configurations on the global level, the prefix "workers." can be used now.

[downstream_first]: KAFKA-13311

Change-Id: Iabb2aa7761d763cb99af03c7040a67c2cce69747