CDPD-29058: Unwire the log4j-appender module
This commit is based on KAFKA-12399 that aims to deprecate the
log4j-appender. The upstream commit rewrites VerifiableLog4jAppender
to use the appender available in log4j2 instead of the old one included
in the Kafka library.

In addition to this we remove log4j-appender as dependency from any
other module but leave it in the pom. This way users depending on it
can still use it, however they won't pull it in via other modules and
therefore won't pull in log4j 1.2.

Change-Id: If3514e53d5ce3d2a94937f0a10c78bfa94dc10e5