KAFKA-14652: Add the flow to the log context and the Connect-managed … (#78)
KAFKA-14652: Add the flow to the log context and the Connect-managed client.ids in MM2 dedicated mode

MM2 runs multiple Connect workers in a single process, and the Connectors inside each worker have identical names. This means that the current logging generates a lot of conflicts, and the logs of the different Connector instances are mixed.
The Connect WorkerConfig is now extended with a contextPrefix method, which provides an optional label for the log context and the client ids. In Connect, it is always null, but in the dedicated MM2 mode, it is set to the flow (source->target).

Change-Id: I6da88130218cef4ec4f710f10f9c2d6b6bd4f965

Co-authored-by: Daniel Urban <durban@cloudera.com>